+++ spec/lib/contentful/management/content_type_spec.rb
@@ -21,6 +21,16 @@ module Contentful
         end
       end
 
+      describe '.all_published' do
+        let!(:space_id) { 'bjwq7b86vgmm' }
+        it 'returns a Contentful::Array' do
+          vcr('content_type/all_public') { expect(subject.all_published(space_id)).to be_kind_of Contentful::Management::Array }
+        end
+        it 'builds a Contentful::Management::ContentType object' do
+          vcr('content_type/all_public') { expect(subject.all_published(space_id).first).to be_kind_of Contentful::Management::ContentType }
+        end
+      end
+
       describe '#find' do
         it 'returns a Contentful::Management::ContentType' do
           vcr('content_type/find') { expect(subject.find(space_id, content_type_id)).to be_kind_of Contentful::Management::ContentType }